ArchiveUtils moved to network package
