+++ feature_book_details/src/main/java/com/allsoftdroid/feature/book_details/presentation/viewModel/BookDetailsViewModel.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.Transformations
 import com.allsoftdroid.common.base.extension.Event
 import com.allsoftdroid.common.base.usecase.BaseUseCase
 import com.allsoftdroid.common.base.usecase.UseCaseHandler
-import com.allsoftdroid.common.base.utils.ArchiveUtils
+import com.allsoftdroid.common.base.network.ArchiveUtils
 import com.allsoftdroid.feature.book_details.domain.model.AudioBookTrackDomainModel
 import com.allsoftdroid.feature.book_details.domain.usecase.GetMetadataUsecase
 import com.allsoftdroid.feature.book_details.domain.usecase.GetTrackListUsecase
